requirements.txt
# SKYT Comprehensive Experiment System Requirements

# =============================================================================
# API Server (FastAPI)
# =============================================================================
fastapi>=0.104.0       # Web framework
uvicorn[standard]>=0.24.0  # ASGI server
pydantic>=2.0.0        # Data validation
python-multipart>=0.0.6  # Form data parsing
python-jose[cryptography]>=3.3.0  # JWT tokens
PyJWT>=2.8.0           # JWT encoding/decoding

# =============================================================================
# Background Workers (Celery)
# =============================================================================
celery>=5.3.0          # Task queue
redis>=5.0.0           # Redis client (broker + cache)
flower>=2.0.0          # Celery monitoring

# =============================================================================
# LLM Providers
# =============================================================================
openai>=1.0.0          # OpenAI GPT models & OpenRouter (uses same SDK)
anthropic>=0.18.0      # Anthropic Claude models

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0

# AST manipulation and code analysis
astor>=0.8.1

# Optional: Enhanced plotting
plotly>=5.0.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.12.0
jsonschema>=4.0.0      # JSON schema validation

# Code quality
black>=21.0.0
flake8>=3.9.0

# Validation framework (Phase 1: Static analysis + Type checking)
radon>=6.0.1       # Complexity metrics (cyclomatic, Halstead, maintainability)
mypy>=1.7.0        # Static type checking and inference
bandit>=1.7.5      # Security vulnerability scanning
